Split impls in helpers.rs into relevant submodules
- Handle-related keys are now in handle.rs
- Property setting/getting is now in src/property.rs
